+++ Cargo.toml
@@ -2,6 +2,8 @@
 members = [
     "tide",
     "tide-compression",
+    "tide-cookies",
+    "tide-core",
     "examples",
 ]
