+++ tide/Cargo.toml
@@ -19,25 +19,17 @@ repository = "https://github.com/rustasync/tide"
 version = "0.2.0"
 
 [dependencies]
-cookie = { version = "0.12", features = ["percent-encode"] }
 futures-preview = "0.3.0-alpha.16"
-fnv = "1.0.6"
 http = "0.1"
 http-service = "0.2.0"
-pin-utils = "0.1.0-alpha.4"
-route-recognizer = "0.1.12"
 serde = "1.0.91"
 serde_derive = "1.0.91"
-serde_json = "1.0.39"
 slog = "2.4.1"
 slog-async = "2.3.0"
 slog-term = "2.4.0"
-typemap = "0.3.3"
 serde_urlencoded = "0.5.5"
-
-[dependencies.http-service-hyper]
-optional = true
-version = "0.2.0"
+tide-cookies = { path = "../tide-cookies", optional = true }
+tide-core = { path = "../tide-core" }
 
 [dependencies.multipart]
 default-features = false
@@ -45,8 +37,9 @@ features = ["server"]
 version = "0.16.1"
 
 [features]
-default = ["hyper"]
-hyper = ["http-service-hyper"]
+default = ["hyper", "cookies"]
+cookies = ["tide-cookies"]
+hyper = ["tide-core/http-service-hyper"]
 
 [dev-dependencies]
 http-service-mock = "0.2.0"
